requirements.txt
# Web framework
Flask==2.3.3
Werkzeug>=2.3.7


# LangChain + Vertex AI integration
langchain>=0.1.159
langchain-google-vertexai>=2.0.0

# Google Cloud SDK
google-cloud-aiplatform>=1.28.0
google-cloud-vision>=3.3.1

# Document processing
PyMuPDF>=1.22.5
docx2txt>=0.8
pdf2image>=1.16.3
pillow>=10.1.0

# NER / Transformers
transformers>=4.42.0
torch>=2.1.0
sentencepiece>=0.1.99
indic-nlp-library>=0.81.6

# Visualization & Graphs
networkx>=3.1
matplotlib>=3.8.0

# HTTP requests
requests>=2.31.0

# Environment management
python-dotenv>=1.0.0

# Numerical processing
numpy>=1.26.0
